refs #4129
git-svn-id: https://devel.scolab.eu/gvsig-online/gvsigonline_v2/trunk/gvsigol@5831 216424d3-d1a8-4ca4-951f-a609719bba38

+++ gvsigol_services/backend_geoserver.py
@@ -326,7 +326,7 @@ class Geoserver():
             print exc
             return False
     
-    def deleteDatastore(self, workspace, datastore, purge=None):
+    def deleteDatastore(self, workspace, datastore, delete_schema, purge=None):
         """
         Deletes a datastore and all its associated resources
         """
@@ -341,7 +341,8 @@ class Geoserver():
                 return True
             else:
                 catalog.delete(ds, purge, recurse=True)
-                utils.delete_schema_for_datastore(json.loads(datastore.connection_params))
+                if delete_schema:
+                    utils.delete_schema_for_datastore(json.loads(datastore.connection_params))
                 
             return True
         except Exception as e:
